Move message into it's own module.
