+++ bitswap/Cargo.toml
@@ -19,3 +19,6 @@ libp2p-swarm = "0.16.1"
 log = "0.4.8"
 prost = "0.6.1"
 thiserror = "1.0.11"
+
+[dev-dependencies]
+multihash = "0.10.1"
