+++ bitswap/src/ledger.rs
@@ -7,8 +7,6 @@ use libipld::cid::{Cid, Prefix};
 use prost::Message as ProstMessage;
 use std::collections::HashMap;
 
-pub type Priority = i32;
-
 /// The Ledger contains the history of transactions with a peer.
 #[derive(Debug, Default)]
 pub struct Ledger {
@@ -70,217 +68,7 @@ impl Ledger {
         }
     }
 }
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct I;
-#[derive(Debug, Clone, PartialEq)]
-pub struct O;
-
-/// A bitswap message.
-#[derive(Clone, PartialEq)]
-pub struct Message<T> {
-    /// Message tag
-    _phantom_data: PhantomData<T>,
-    /// List of wanted blocks.
-    want: HashMap<Cid, Priority>,
-    /// List of blocks to cancel.
-    cancel: Vec<Cid>,
-    /// Wheather it is the full list of wanted blocks.
-    full: bool,
-    /// List of blocks to send.
-    blocks: Vec<Block>,
-}
-
-impl<T> Default for Message<T> {
-    fn default() -> Self {
-        Message {
-            _phantom_data: PhantomData,
-            want: HashMap::new(),
-            cancel: Vec::new(),
-            full: false,
-            blocks: Vec::new(),
-        }
-    }
-}
-
-impl<T> Message<T> {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Returns the list of blocks.
-    pub fn blocks(&self) -> &[Block] {
-        &self.blocks
-    }
-
-    /// Returns the list of wanted blocks.
-    pub fn want(&self) -> &HashMap<Cid, Priority> {
-        &self.want
-    }
-
-    /// Returns the list of cancelled blocks.
-    pub fn cancel(&self) -> &[Cid] {
-        &self.cancel
-    }
-
-    /// Adds a `Block` to the message.
-    pub fn add_block(&mut self, block: Block) {
-        self.blocks.push(block);
-    }
-
-    /// Removes the block from the message.
-    pub fn remove_block(&mut self, cid: &Cid) {
-        self.blocks.retain(|block| block.cid() != cid);
-    }
-
-    /// Adds a block to the want list.
-    pub fn want_block(&mut self, cid: &Cid, priority: Priority) {
-        self.want.insert(cid.to_owned(), priority);
-    }
-
-    /// Adds a block to the cancel list.
-    pub fn cancel_block(&mut self, cid: &Cid) {
-        self.cancel.push(cid.to_owned());
-    }
-
-    /// Removes the block from the want list.
-    #[allow(unused)]
-    pub fn remove_want_block(&mut self, cid: &Cid) {
-        self.want.remove(cid);
-    }
-}
-
-impl Into<Vec<u8>> for &Message<O> {
-    fn into(self) -> Vec<u8> {
-        let mut proto = bitswap_pb::Message::default();
-        let mut wantlist = bitswap_pb::message::Wantlist::default();
-        for (cid, priority) in self.want() {
-            let mut entry = bitswap_pb::message::wantlist::Entry::default();
-            entry.block = cid.to_bytes();
-            entry.priority = *priority;
-            wantlist.entries.push(entry);
-        }
-        for cid in self.cancel() {
-            let mut entry = bitswap_pb::message::wantlist::Entry::default();
-            entry.block = cid.to_bytes();
-            entry.cancel = true;
-            wantlist.entries.push(entry);
-        }
-        for block in self.blocks() {
-            let mut payload = bitswap_pb::message::Block::default();
-            payload.prefix = block.cid().prefix().as_bytes();
-            payload.data = block.data().to_vec();
-            proto.payload.push(payload);
-        }
-        if !wantlist.entries.is_empty() {
-            proto.wantlist = Some(wantlist);
-        }
-        let mut res = Vec::with_capacity(proto.encoded_len());
-        proto
-            .encode(&mut res)
-            .expect("there is no situation in which the protobuf message can be invalid");
-        res
-    }
-}
-
-impl Message<O> {
-    /// Turns this `Message` into a message that can be sent to a substream.
-    pub fn to_bytes(&self) -> Vec<u8> {
-        self.into()
-    }
-}
-
-impl TryFrom<&[u8]> for Message<I> {
-    type Error = BitswapError;
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        let proto: bitswap_pb::Message = bitswap_pb::Message::decode(bytes)?;
-        let mut message = Message::new();
-        for entry in proto.wantlist.unwrap_or_default().entries {
-            let cid = Cid::try_from(entry.block)?;
-            if entry.cancel {
-                message.cancel_block(&cid);
-            } else {
-                message.want_block(&cid, entry.priority);
-            }
-        }
-        for payload in proto.payload {
-            let prefix = Prefix::new_from_bytes(&payload.prefix)?;
-            let cid = Cid::new_from_prefix(&prefix, &payload.data);
-            let block = Block {
-                cid,
-                data: payload.data.to_vec().into_boxed_slice(),
-            };
-            message.add_block(block);
-        }
-        Ok(message)
-    }
-}
-
-impl Message<I> {
-    /// Creates a `Message` from bytes that were received from a substream.
-    pub fn from_bytes(bytes: &[u8]) -> Result<Self, BitswapError> {
-        Self::try_from(bytes)
-    }
-}
-
-impl<T> std::fmt::Debug for Message<T> {
-    fn fmt(&self, fmt: &mut std::fmt::Formatter) -> Result<(), std::fmt::Error> {
-        for (cid, priority) in self.want() {
-            writeln!(fmt, "want: {} {}", cid.to_string(), priority)?;
-        }
-        for cid in self.cancel() {
-            writeln!(fmt, "cancel: {}", cid.to_string())?;
-        }
-        for block in self.blocks() {
-            writeln!(fmt, "block: {}", block.cid().to_string())?;
-        }
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    /*
-    use super::*;
-
-    #[test]
-    fn test_empty_message_to_from_bytes() {
-        let message = Message::new();
-        let bytes = message.clone().into_bytes();
-        let new_message = Message::from_bytes(&bytes).unwrap();
-        assert_eq!(message, new_message);
-    }
-
-    #[test]
-    fn test_want_message_to_from_bytes() {
-        let mut message = Message::new();
-        let block = Block::from("hello world");
-        message.want_block(&block.cid(), 1);
-        let bytes = message.clone().into_bytes();
-        let new_message = Message::from_bytes(&bytes).unwrap();
-        assert_eq!(message, new_message);
-    }
-
-    #[test]
-    fn test_cancel_message_to_from_bytes() {
-        let mut message = Message::new();
-        let block = Block::from("hello world");
-        message.cancel_block(&block.cid());
-        let bytes = message.clone().into_bytes();
-        let new_message = Message::from_bytes(&bytes).unwrap();
-        assert_eq!(message, new_message);
-    }
-
-    #[test]
-    fn test_payload_message_to_from_bytes() {
-        let mut message = Message::new();
-        let block = Block::from("hello world");
-        message.add_block(block);
-        let bytes = message.clone().into_bytes();
-        let new_message = Message::from_bytes(&bytes).unwrap();
-        assert_eq!(message, new_message);
-    }
-
+/*
     #[test]
     fn test_ledger_send_block() {
         let block_1 = Block::from("1");
